all: migrate to using sqlc
We want to get out of the business of generating SQL commands
altogether because it's fraught with peril. Instead have sqlc generate
them automatically from the queries.

There's more to do here including threading through
a `context.Context` everywhere but this is a good start.

Update the lock Acquire strategy to use advisory locks instead of
trying to directly update. Some small benchmarks indicate that
performance is better using this strategy. We are in the range of
hundreds of jobs dequeued per second which is honestly fine for a
project like this; if this is insufficient you can use any of the
other high performance tools, or shard to multiple databases, or any
other tool.

+++ downstream/job.go
@@ -15,7 +15,7 @@ type JobService struct {
 
 type JobParams struct {
 	Data     json.RawMessage `json:"data"`
-	Attempts uint8           `json:"attempts"`
+	Attempts int16           `json:"attempts"`
 }
 
 // Post makes a request to /v1/jobs/:job-name/:job-id with the job data.
